+++ src/scip/reader_cnf.c
@@ -160,7 +160,7 @@ SCIP_RETCODE readCnf(
    linecount = 0;
 
    /* read header */
-   SCIP_CALL( readCnfLine(scip, file, line, sizeof(line), &linecount) );
+   SCIP_CALL( readCnfLine(scip, file, line, (int) sizeof(line), &linecount) );
    if( *line != 'p' )
    {
       readError(scip, linecount, "problem declaration line expected");
@@ -216,7 +216,7 @@ SCIP_RETCODE readCnf(
    clauselen = 0;
    do
    {
-      retcode = readCnfLine(scip, file, line, sizeof(line), &linecount);
+      retcode = readCnfLine(scip, file, line, (int) sizeof(line), &linecount);
       if( retcode != SCIP_OKAY )
          goto TERMINATE;
